v1.0 - refactor dan bugfixing: done

+++ src/appbase/arins/Traits/Http/Controller/Post/Actionresponse.php
@@ -5,16 +5,11 @@ namespace Arins\Traits\Http\Controller\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
-use Arins\Facades\Response;
-use Arins\Facades\Filex;
-use Arins\Facades\Formater;
-use Arins\Facades\ConvertDate;
-
-trait Init
+trait Actionresponse
 {
 
     /** post */
-    public function initStore(Request $request)
+    public function store(Request $request)
     {
         //get input value by fillable fields
         $data = $request->only($this->data->getFillable()); //get field input
@@ -59,7 +54,7 @@ trait Init
     }
 
     /** post */
-    public function initUpdate(Request $request, $id)
+    public function update(Request $request, $id)
     {
         //get data from database
         $record = $this->data->find($id);
@@ -70,6 +65,7 @@ trait Init
         $upload = $request->file('upload'); //upload file (image/document) ==> if included
         $imageTemp = $request->input('imageTemp'); //temporary file uploaded
         $toggleRemoveImage = $request->input('toggleRemoveImage');
+        //return dd($toggleRemoveImage);
 
         // return dd($data);
         // //convert input value (string/date/number/email/etc)
@@ -83,18 +79,35 @@ trait Init
         //validate input value
         $request->validate($this->data->getValidateInput());
 
-        //copy temporary uploaded image to real path
-        $data['image'] = Filex::uploadOrCopyAndRemove('', $uploadTemp, 'activities', $upload, 'public', false);
-        // Filex::delete($imageOld);
+        $imageNew = Filex::uploadOrCopyAndRemove($imageOld, $uploadTemp, 'activities', $upload, 'public', false);
+        $data['image'] = $imageNew;
+        if (strtolower($toggleRemoveImage) ==  'true')
+        {
+            $data['image'] = null;
+        }
 
         if ($this->data->update($record, $data)) {
-            Filex::delete($imageOld);
+            if ($uploadTemp != null)
+            {
+                Filex::delete($imageOld);
+                Filex::delete($uploadTemp);
+            } //end if
+
+            if (strtolower($toggleRemoveImage) == 'true')
+            {
+                Filex::delete($imageOld);
+                Filex::delete($imageNew);
+                Filex::delete($uploadTemp);
+            }
             return redirect()->route($this->sViewName . '.index');
         }
 
         /** jika tetap terjadi kesalahan maka ada kesalahan pada system */
         //step 1: delete image if fail to save
-        Filex::delete($data['image']);
+        if ($uploadTemp != null)
+        {
+            Filex::delete($data['image']);
+        } //end if
 
         //step 2: Kembali ke halaman input
         return redirect()->route($this->sViewName . '.edit', $id)
@@ -102,7 +115,7 @@ trait Init
     }
 
     /** post */
-    public function initDestroy($id)
+    public function destroy($id)
     {
         //
         $record = $this->data->find($id);
